Enable `pg_dump` functionality.
Provide SQL adapters that allow `pg_dump` to capture the contents of an
EdgeDB database. This functionality is not meant to be used as backup
(it will not be sufficient to restore the original EdgeDB database, use
`edgedb dump` tools for that). It is instead meant to be used as a way
to transfer data to a Postgres database potentially as a compatibility
layer with other SQL tooling.